套装奖励tooltip

+++ src/main/java/org/confluence/lib/util/LibUtils.java
@@ -309,8 +309,8 @@ public final class LibUtils {
         }
     }
 
-    public static boolean checkChance(double value, RandomSource random) {
-        return value >= 1.0 || (value > 0.0 && random.nextDouble() < value);
+    public static boolean checkChance(float value, RandomSource random) {
+        return value >= 1.0F || (value > 0.0F && random.nextFloat() < value);
     }
 
     public static <K, V> Map<K, V> convertTupleListToMap(List<Tuple<K, V>> list) {
